Join termination calls

+++ src/vs/workbench/services/extensions/electron-browser/localProcessExtensionHost.ts
@@ -238,6 +238,7 @@ export class LocalProcessExtensionHost implements IExtensionHost {
 
 				// Run Extension Host as fork of current process
 				this._extensionHostProcess = fork(FileAccess.asFileUri('bootstrap-fork', require).fsPath, ['--type=extensionHost'], opts);
+				this._extensionHostProcess.unref();
 
 				// Catch all output coming from the extension host process
 				type Output = { data: string, format: string[] };
@@ -528,6 +529,7 @@ export class LocalProcessExtensionHost implements IExtensionHost {
 	private _onExtHostProcessExit(code: number, signal: string): void {
 		if (this._terminating) {
 			// Expected termination path (we asked the process to terminate)
+			this._extensionHostProcess?.kill();
 			return;
 		}
 
@@ -607,30 +609,31 @@ export class LocalProcessExtensionHost implements IExtensionHost {
 		return withNullAsUndefined(this._inspectPort);
 	}
 
-	public terminate(): void {
-		if (this._terminating) {
-			return;
-		}
-		this._terminating = true;
+	public terminate(): Promise<void> {
+		return new Promise((c) => {
+			if (this._terminating) {
+				c();
+			}
+			this._terminating = true;
 
-		this._toDispose.dispose();
+			this._toDispose.dispose();
 
-		if (!this._messageProtocol) {
-			// .start() was not called
-			return;
-		}
-
-		this._messageProtocol.then((protocol) => {
-			// Send the extension host a request to terminate itself
-			// (graceful termination)
-			protocol.send(createMessageOfType(MessageType.Terminate));
-			protocol.dispose();
-		}).finally(() => {
-			this._cleanResources();
+			if (this._messageProtocol) {
+				this._messageProtocol.then((protocol) => {
+					// Send the extension host a request to terminate itself
+					// (graceful termination)
+					protocol.send(createMessageOfType(MessageType.Terminate));
+					protocol.flush();
+					protocol.dispose();
+					c(this._cleanResources(false));
+				}).catch(() => {
+					c(this._cleanResources(true));
+				});
+			}
 		});
 	}
 
-	private _cleanResources(): void {
+	private _cleanResources(killChild: boolean): void {
 		if (this._namedPipeServer) {
 			this._namedPipeServer.close();
 			this._namedPipeServer = null;
@@ -640,8 +643,10 @@ export class LocalProcessExtensionHost implements IExtensionHost {
 			this._extensionHostConnection = null;
 		}
 		if (this._extensionHostProcess) {
-			this._extensionHostProcess.kill();
-			this._extensionHostProcess = null;
+			if (killChild) {
+				this._extensionHostProcess.kill();
+				this._extensionHostProcess = null;
+			}
 		}
 	}
 
@@ -653,6 +658,6 @@ export class LocalProcessExtensionHost implements IExtensionHost {
 			this._extensionHostDebugService.terminateSession(this._environmentService.debugExtensionHost.debugId);
 			event.join(timeout(100 /* wait a bit for IPC to get delivered */), 'join.extensionDevelopment');
 		}
-		this.terminate();
+		event.join(this.terminate(), 'join.extensionHostTerminate');
 	}
 }
